refactor: cut  all Dao about Repair to dao directory
