+++ src/main/java/org/example/techcare/dao/RepairDetailsDAO.java
@@ -1,7 +1,9 @@
-package org.example.techcare.model.repair;
+package org.example.techcare.dao;
 
 import org.example.techcare.model.component.Component;
 import org.example.techcare.model.component.ComponentDAO;
+import org.example.techcare.model.repair.Repair;
+import org.example.techcare.model.repair.RepairDetails;
 import org.example.techcare.model.utils.ConnectionBdd;
 
 import java.sql.PreparedStatement;
